+++ src/app/style/community-style.css
@@ -1,0 +1,54 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+@layer base {
+  .btn {
+    margin-top: 1px;
+    padding: 2px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    width: 100%;
+  }
+  .input {
+    background-color: transparent;
+    border: none;
+    border-bottom: 1px solid #ccc;
+    color: #555;
+    box-sizing: border-box;
+    font-family: 'Arvo';
+    font-size: 18px;
+    width: 60%;
+    text-align: center;
+    padding-bottom: 4px;
+    &::focus {
+      outline: none;
+    }
+  }
+
+  input::-webkit-input-placeholder {
+    color: #aaa;
+  }
+
+  input:focus::-webkit-input-placeholder {
+    color: #a0d683;
+  }
+
+  .input:focus + .underline {
+    transform: scale(1);
+  }
+  .underline {
+    background-color: #a0d683;
+    display: inline-block;
+    height: 2px;
+    left: 0;
+    margin-top: -4px;
+    position: absolute;
+    bottom: 2px;
+    -webkit-transform: scale(0, 1);
+    transform: scale(0, 1);
+    -webkit-transition: all 0.5s linear;
+    transition: all 0.5s linear;
+    width: 60%;
+  }
+}
